creat the fine prodect works

+++ src/app/Copmonets/user/user.ts
@@ -1,7 +1,7 @@
 import { User as UserService } from './../../services/user';
 import { Component, signal } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import id from '@angular/common/locales/id';
+import { Console } from 'console';
 
 @Component({
   selector: 'app-user',
@@ -38,20 +38,23 @@ export class User {
         this.isloading.set(false);
       }
     );
-    //delete user
   }
-  deleteUser(id: number) {
-    this.isloading.set(true);
-    this.userService.deleteUser(id).subscribe(
-      (res: any) => {
-        const updatedUsers = this.users().filter((user) => user.id !== id);
-        this.users.set(updatedUsers);
-        this.isloading.set(false);
-      },
-      (error) => {
-        console.error('Error deleting user:', error);
-        this.isloading.set(false);
-      }
-    );
+
+  deleteUser(id: any) {
+    if (confirm('Are you sure you want to delete this user?')) {
+      this.isloading.set(true);
+      this.userService.deleteUser(id).subscribe(
+        (res: any) => {
+          const updatedUsers = this.users().filter((user) => user.id !== id);
+          this.users.set(updatedUsers);
+          this.isloading.set(false);
+          console.log('Deleted user response:', res);
+        },
+        (error) => {
+          console.error('Error deleting user:', error);
+          this.isloading.set(false);
+        }
+      );
+    }
   }
 }
